requirements.txt
nvidia-cublas-cu12==12.4.5.8
xformers==0.0.27
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
easydict==1.13
einops==0.8.1
peft==0.14.0
pillow==11.0.0
protobuf==5.29.3
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
transformers==4.49.0
datasets==3.6.0
scikit-learn==1.6.1
moviepy==1.0.3
imageio==2.33.1
imageio-ffmpeg==0.4.9
opencv-python==4.11.0.86
diffusers==0.32.2
wandb
matplotlib
lpips
scikit-image